[thread-store] use open-ended item update scans
