+++ codex-rs/thread-store/src/lib.rs
@@ -6,6 +6,7 @@
 
 mod error;
 mod in_memory;
+mod list_items;
 mod live_thread;
 mod local;
 mod store;
@@ -16,6 +17,11 @@ pub use error::ThreadStoreError;
 pub use error::ThreadStoreResult;
 pub use in_memory::InMemoryThreadStore;
 pub use in_memory::InMemoryThreadStoreCalls;
+pub use list_items::ItemPage;
+pub use list_items::ListItemsFilters;
+pub use list_items::ListItemsParams;
+pub use list_items::StoredThreadItem;
+pub use list_items::UpdatedAtFilter;
 pub use live_thread::LiveThread;
 pub use live_thread::LiveThreadInitGuard;
 pub use local::LocalThreadStore;
@@ -29,9 +35,6 @@ pub use types::CreateThreadParams;
 pub use types::DeleteThreadParams;
 pub use types::ExtraConfig;
 pub use types::GitInfoPatch;
-pub use types::ItemPage;
-pub use types::ListItemsFilters;
-pub use types::ListItemsParams;
 pub use types::ListThreadsParams;
 pub use types::ListTurnsParams;
 pub use types::LoadThreadHistoryParams;
@@ -42,7 +45,6 @@ pub use types::SearchThreadsParams;
 pub use types::SortDirection;
 pub use types::StoredThread;
 pub use types::StoredThreadHistory;
-pub use types::StoredThreadItem;
 pub use types::StoredThreadSearchResult;
 pub use types::StoredTurn;
 pub use types::StoredTurnError;
@@ -56,4 +58,3 @@ pub use types::ThreadSearchPage;
 pub use types::ThreadSortKey;
 pub use types::TurnPage;
 pub use types::UpdateThreadMetadataParams;
-pub use types::UpdatedAtFilter;
