+++ codex-rs/thread-store/src/types.rs
@@ -372,73 +372,6 @@ pub struct TurnPage {
     pub backwards_cursor: Option<String>,
 }
 
-/// Optional filters for listing persisted items.
-#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ListItemsFilters {
-    /// Optional turn id to filter by. When omitted, returns items across the thread.
-    pub turn_id: Option<String>,
-
-    /// Optional item update timestamp filter.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub updated_at: Option<UpdatedAtFilter>,
-}
-
-/// Bounds for filtering item snapshots by their update timestamp.
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct UpdatedAtFilter {
-    /// Exclusive lower bound: `updated_at > gt`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub gt: Option<DateTime<Utc>>,
-
-    /// Inclusive upper bound: `updated_at <= lte`.
-    pub lte: DateTime<Utc>,
-}
-
-/// Parameters for listing persisted items within a thread.
-///
-/// Callers must reuse the same filters when following a page cursor. Implementations that forward
-/// an [`UpdatedAtFilter`] must verify that the backing store applied its exact `lte` bound before
-/// returning a successful page.
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ListItemsParams {
-    /// Thread id to read.
-    pub thread_id: ThreadId,
-    /// Whether archived threads are eligible.
-    pub include_archived: bool,
-    /// Filters applied to the listed items.
-    ///
-    /// Flattening preserves the existing serialized `turn_id` field.
-    #[serde(flatten)]
-    pub filters: ListItemsFilters,
-    /// Opaque cursor returned by a previous list call.
-    pub cursor: Option<String>,
-    /// Maximum number of items to return.
-    pub page_size: usize,
-    /// Sort direction requested by the caller.
-    pub sort_direction: SortDirection,
-}
-
-/// A projected app-server `ThreadItem` snapshot within a turn.
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct StoredThreadItem {
-    pub turn_id: Option<String>,
-    pub item_key: String,
-    pub item_ordinal: u64,
-    pub item_created_at_ms: i64,
-    pub materialized_thread_item_json: Vec<u8>,
-}
-
-/// A page of persisted items within a thread, optionally filtered to a turn.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct ItemPage {
-    /// Items returned for this page.
-    pub items: Vec<StoredThreadItem>,
-    /// Opaque cursor to continue listing.
-    pub next_cursor: Option<String>,
-    /// Opaque cursor for fetching in the opposite direction.
-    pub backwards_cursor: Option<String>,
-}
-
 /// Store-owned thread metadata used by list/read/resume responses.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct StoredThread {
@@ -765,30 +698,6 @@ mod tests {
     use serde_json::json;
 
     use super::*;
-
-    #[test]
-    fn list_items_filters_preserve_flattened_turn_id_shape() {
-        let params = ListItemsParams {
-            thread_id: ThreadId::default(),
-            include_archived: true,
-            filters: ListItemsFilters {
-                turn_id: Some("turn-1".to_string()),
-                updated_at: None,
-            },
-            cursor: Some("next-page".to_string()),
-            page_size: 25,
-            sort_direction: SortDirection::Asc,
-        };
-
-        let value = serde_json::to_value(&params).expect("serialize list items params");
-        assert_eq!(value["turn_id"], json!("turn-1"));
-        assert_eq!(value.get("filters"), None);
-        assert_eq!(value.get("updated_at"), None);
-
-        let decoded: ListItemsParams =
-            serde_json::from_value(value).expect("deserialize legacy-shaped list items params");
-        assert_eq!(decoded, params);
-    }
 
     #[test]
     fn thread_metadata_patch_round_trips_optional_clears() {
